New option 'full-search' to control copy-mode history search coverage
New option 'full-search' to control if copy-mode will search
over the full history.  Defaults to on, when off number of
results and highlighting is not shown.

When dealing with a history of ~20000 lines there's a small,
but noticable, lag in the amouot of time to jump to the first
result and search-again.  >50k+ histories results in multi-second wait.