Import OpenSSH's setproctitle() compat shim.
tmux currently uses prctl(2) on Linux in order to change the process
title, which permits a thread to set its process title to a given 16
byte string which is then exposed in /proc/<PID>/comm. However this
means that the changes to process title are not visible to programs such
as htop(1), which instead read /proc/<PID>/cmdline.

The way to expose such information in the cmdline file (and what
OpenSSH's compat shim does) is to make a copy of argv and environ and
then overwrite the original argv/environ space passed to main().

OpenSSH also has support for systems where a similar effect is achieved
using a pstat() function, which is included in this commit.