copy-{selection,pipe}: reuse code between variants
copy-selection, copy-selection-no-clear and copy-selection-and-cancel
are identical except their return value and whether or not they clear
the selection. Same for copy-pipe.

Reuse the code between the variants with the -no-clear variant as base.

Note that copy-pipe-and-cancel previously had two potential return
values: WINDOW_COPY_CMD_NOTHING and ..CMD_CANCEL, depending on whether
or not there was a selection, and now it always returns ..CMD_CANCEL .

This is now analogous to copy-selection-and-cancel which was and still
is returning ..CMD_CANCEL unconditionally.