Remove use of REG_STARTEND
The REG_STARTEND flag is not a part of POSIX standard regular
expressions (see [1]). Avoid using it to maintain tmux's portability to
many operating systems.

Removing it is fairly straightforward. We never used it to limit the
search before the end of the string, we only used it to move search start
offset forward. Instead of passing REG_STARTEND, simply pass string +
start into regexec(). Fix up uses of rm_so and rm_eo to be relative to
start, and Bob's your uncle.

[1] https://pubs.opengroup.org/onlinepubs/009695399/functions/regexec.html