Add a highlight-search option
Highlight-search can make searching unusable on really large buffers. This is due to it trying to look for all occurrences of the result in the whole buffer, which can be very slow. We cannot simply highlight the occurrences in the current view as that requires updating the search as the view changes and while this is possible seemed very complex.

To repro:
set -g history-limit 300000

then run:
$ seq 300000

Searching through this buffer is very sluggish due to the search algorithm.

With highlight-search set to 0 this works much better.