escaping now works with styles
Styles are interpreted in format_draw instead of format_expand1, so the
escaping octothorp must make it to format_draw for it to know this isn't
a style.  This also means format_draw must handle removing the extra
octothorp before displaying and take them into account during
calculations for width and left/right trimming.