(feature): Add client-detached message to control-mode
Closes #2598

This commit adds an extra message to command-mode to indicate when a
client has detached from tmux.

I wasn't sure whether there was any need to distinguish between the
current client detaching and an external client detaching (such as the
session changes message has both %session-changed for the current command-mode
client and %client-session-changed for another client doing so) so this
commit doesn't distinguish between the current and another client.