Change special key code constants to macros instead of enums
The C standard specifies that enumeration values must fit within a
signed integer.  However, it is implementation-defined what type the
compiler may choose.  For the Sun Studio compiler, that type is `int`
whereas `gcc` defaults to `unsigned int`.

This change avoids value overflow by changing the constants to be
defined as macros.