Fix a regression where client doesn't exit when stdout write fails
When the output of tmux client is pipelined and the latter process fails
without reading all the output, tmux is supposed to terminate silently.

Due to the changes introduced by the commit e40831a, tmux client fails
to exit when writes to stdout fails. `client_write_error_callback`
correctly cleans up the event relevant to the client file for stdout. But
the problem is that exit-checking code `client_file_check_cb` is called
before the clean-up. As the result, it fails to recognize the write failure
and the tmux client remains in the event loop and the tmux server does
the same.

This patch moves the call for `client_file_check_cb` after the clean-up
in `file_write_error_callback`, which is the same to the code before e40831a.