Improve error message if socket dir has too liberal permissions
tmux expects that the directory containing the socket used to
communicate with the server process to be owned by the user and that
others have no permission to access it.

If this fails however, the emitted error message is misleading as the
directory exists just fine:

	uwe@taurus:~$ chmod o+x /tmp/tmux-1000
	uwe@taurus:~$ tmux a
	error creating /tmp/tmux-1000 (Permission denied)

Rework error handling in make_label() to emit an error message that
talks about unsafe permissions in this case instead.

Debian-Bug: https://bugs.debian.org/990215
Fixes: f32fd2df69d4 ("Improve error message if creating socket parent directory fails [...]")