tmux.1: Version specific configuration example
While there are many ways how one can make a config file compatible with
multiple versions of tmux, it would be nice to provide an example most
people can refer to and copy from.

It would also document an "official" way of how the "multi-version
config files" should be written, stating support for this approach in
the future.