Fix handling of ST terminator when parsing OSC-52 reply
While parsing an OSC-52 reply, we search the input buffer for both BEL
and ST.

Once one of them has been found, we proceed to copy and base64 decode
the payload, and then adding the decoded data to a new paste buffer.

The check for ST looks backward in the buffer. I.e. it checks if
the *previous* character is ESC, and the *current* character is \.

This means, our ‘end’ index points to the *end* of the
terminator. This is correct when calculating the length of the entire
escape sequence, which is done immediately after having found the
terminator.

It is however *not* correct when copying out the payload. When an
OSC-52 reply is ST terminated, we end up including ESC in the
payload. This in turn causes the base64 decode to fail, and no paste
buffer is created.