treat C1 control characters as being zero width
wcwidth uses -1 to mean both "this character is not printable" and "this
character is unknown", which means that always using a default of 1
gives incorrect results for control characters (they are not printable,
so they should have a width of 0, not 1). we don't have to care about C0
control characters here, because they are never treated as text, but C1
control characters are, and so we should use the correct values for
them.

this is only an issue with wcwidth - utf8proc returns useful values here
already.