create list of last visited panes for each window
This commit creates a stack-like list of the last visited panes for each
window so that the user can use the "last-pane" command to visit
previously seen panes even if the most recently seen pane(s) are
deleted.

It also adjusts the winlink stack to use a "visited" flag so that it
does not need to loop over the entire stack when removing winlinks.