Use DEC mode 2026 for Sync terminal feature
The ESC P = %p1 s ST sequence is an iTerm2 extension to enable or
disable synchronized output. However, some effort has been made recently
to "standardize" (as much as anything can be standardized in this world)
on using DEC mode 2026 for this feature instead [1]. iTerm2 also
supports this mode and even recommends its use over its own extension
[2].

As far as I can tell the only other terminal emulator that uses iTerm2's
extension is kitty, but they copied it from tmux. So if tmux changes to
use DEC mode 2026 then that would make a strong case for kitty to change
as well.

[1]: https://gist.github.com/christianparpart/d8a62cc1ab659194337d73e399004036
[2]: https://gitlab.com/gnachman/iterm2/-/wikis/synchronized-updates-spec#decsetdecreset