Copy command flags for clipboard and paste buffers
This resolves issue #4123 making it possible to pass arguments to all
the copy commands (i.e. copy-selection, copy-pipe, copy-line, etc) to
control where the copied text is sent. The default behaviour remains the
same, however, adding the '-C' option prevents the commands from sending
the text to the clipboard, and adding the '-P' option prevents them from
adding the text to the paste buffers.