Fix commands not expecting flags
There are certain commands in window-copy.c that expect to get raw text
as their first argument. Now that the argument handling for these
commands has been overhauled, an initial text argument starting with a
dash may cause the argument parser to decide the command invocation has
an invalid flag and ignore it entirely. Instead, these commands should
now precede the raw text argument with a '--' to ensure the argument
parser knows to treat it like text and not a flag.

This commit updates all invocations of the affected commands inside the
tmux repository to ensure they work as before.