Stopped ignoring error on client command parse.
I figure this might be worth doing. I know now we output the error on
the client and the server now has error handling code for nothing in
it's argument handling code. I think ideally we should pass the parsed
cmd structure to the server and thus avoid parsing twice. This commit
doesn't do that yet.