Make it easy to swap windows from choose-tree
I picked the new keys for this because of the similarity to switching vs
swapping tabs in a browser or other tabbed UI. There, Ctrl+PgUp switches
to one tab previous and Ctrl+Shift+PgUp swaps with the previous tab. So
I figured adding Shift to swap instead of move might be somewhat natural
here.

Some of the swapping code is copied from cmd-swap-window.c

Fixes #4388