Improve documentation on the type of options
To be more close to the implementation, use 'scope' for what was
previously called 'type', and improve the documentation on the actual
option types (i.e. flag, choices, ...).

Rephrase the sentence before {set,show}-options commands to be similar
to the one in the HOOKS section, since show-options does not actually
set options.

Co-authored-by: Julian Prein <julian@druck.dev>