Detect support for OSC 52 using the device attributes report
This is an extension agreed upon by modern terminals to indicate that
they support copying to the clipboard with XTerm's OSC 52 sequence.
On terminals that can disable clipboard access, this extension is only
reported when writing to the clipboard is actually allowed, so this is
a more reliable indicator than the Ms terminfo capability.