Fix integer overflow in `grid_clear` arguments
`grid_view_delete_lines` and `grid_view_delete_cells` use wrong logic to
calculate the number of lines or cells to be cleared.

For `grid_view_delete_lines`, lines after the deleted area are moved by `ny`
lines (the number of lines to be deleted), so the remaining number of lines to
be cleared is `ny`.

Likewise, `grid_view_delete_cells` moves cells by `nx`, so final `nx` cells
need to be cleared.

Closes #4598.