Fix OOB access in sixel printing
Fixes OOB access in sixel_print by ensuring that the number of chunks
actually matches the highest color used on the image, even if not defined as
a color register.

We still pass those color numbers through, as such images are perfectly
valid - indeed, on the VT340, all images shared the same color registers,
so not defining any at all is not unheard of.

(I first thought this was my mistake, but at a second glance it looks like
the precursor to the chunks/active variables had the same issue.)